Move DisplayableItem to feature/common
