+++ app/src/main/java/com/futurice/freesound/feature/search/SearchFragmentModule.java
@@ -16,12 +16,12 @@
 
 package com.futurice.freesound.feature.search;
 
+import com.futurice.freesound.feature.common.DisplayableItem;
 import com.futurice.freesound.feature.common.Navigator;
 import com.futurice.freesound.inject.fragment.BaseFragmentModule;
 import com.futurice.freesound.inject.fragment.FragmentScope;
 import com.futurice.freesound.ui.adapter.AdapterInteractor;
 import com.futurice.freesound.ui.adapter.base.DefaultAdapterInteractor;
-import com.futurice.freesound.viewmodel.DisplayableItem;
 import com.squareup.picasso.Picasso;
 
 import dagger.Module;
